Fix some names, add dedicated Trie modules, fix hints, remove error and undefined
All of the main BCP47 tag types are called BCP47 (requiring
qualification if they are to be used simultaneously, but their modules
were probably going to require qualification anyway). Some functions
related to rendering subtags now have Lower attached to them, in case
we ever expand the scope of Subtag to include different characters.

The Trie things now live in their own modules. Eventually they should
be developed further, but the library doesn't need them right now.

The last use of error was removed, which would only occur in one
function if a Subtag invariant were violated. I prefer to have this
return "" instead of raising the error. The last undefined (other than
the one related to Unbox) was also removed, which required the
wrapSubtag function to be fleshed out.

The error function is still used heavily in languagetags-gen, but that
is less of an issue at the moment.