Move Subtag* to BCP47.Subtag*, add generated function to RedundantRecords
It's easiest for backwards-compatibility if I simply have Subtag live
in BCP47. That way, if we need to consider tags with upper case
characters or symbols we can simply create the a new Subtag module and
not have that act be a breaking change.

RedundantRecords now has a function that converts a redundant tag to
its corresponding valid tag