Rework Subtag and Syntax parsing, remove or rename certain functions
The grandfatheredToTag function is now grandfatheredSyntax, so that
Syntax can merely export the one from Registry.Grandfathered rather
than defining it itself.

The parseSubtagMangled and packCharMangled functions were removed, as
were the related tag parsing functions from Syntax. These really
shouldn't have existed.

Subtag.SyntaxError now has an NFData instance

The EmptySubtag constructor was added to both the Subtag.SyntaxError
and Syntax.SyntaxError types. These are thrown when a "--" sequence is
encountered, or when the input starts with "-".

The Syntax.TrailingTerminator constructor is now drastically
simplified, since it ought only ever to be thrown when the input
consists exactly of a well-formed tag followed by a dash. This change
required a moderate re-working of the parsing in Syntax, notably
requiring tagPop to accept continuations so that it could handle
trailing dashes itself. The end result is a slightly cleaner parser.

Irregular grandfathered tags not starting with "i" that occur as
strict prefixes of the input are now properly detected. The result is
a slightly better error in those (admittedly very niche) cases.

Various unneeded type signatures were removed.

The parseSubtag function was rewritten to use popSubtag internally,
since changes to those functions have resulted in substantial
convergence in their implementations. More tests were added to ensure
that popSubtag and parseSubtag remain correct.

A few additional unit tests and one property test were added to
SyntaxSpec (related to the SyntaxError changes). More still need to be
added.