Make Quasi errors nicer, add rendering and wrapping, slight renaming
The errors in Quasi now describe exactly what is wrong with the tags
or subtags they are given. To do this, syntax error rendering and
helper functions were added to both Subtag and Syntax. Some tests for
the quasi-quotes were added to a new QuasiSpec module, though without
a template haskell testing library these are somewhat limited.

The registered tag and subtag modules now export rendering functions
for their types.

The wrapChar function was added in analogy with wrapSubtag.

ValidateError was renamed to ValidationError. Its constructors were
also renamed.