Overhaul Trie, general clean-up, add bounds to cabal
The languagetags-gen executable can now be run either from the project
root or the ./languagetags-gen subdirectory, via a very primitive
location check.

All of the Registry.*Records modules are now exported, so that we can
avoid exporting the details vectors from the public-facing modules.
These modules now have descriptions as well.

A .hlint.yaml file was added, which suppresses certain hlint warnings
in one of the test modules. The other hlint suggestions were fixed.

The package description was changed slightly.

Bounds were added to all of the dependencies in the package. The
lowest versions that seem to work all roughly date to GHC 8.6.5 and
base-4.12. We could probably have this package work with earlier
compilers, but that would be a bit of a pain.

The Quasi module is now compatible with the latest template-haskell. A
small compatibility function needed to be created to work around a
change in the TupE tuple expression constructor.

The Validation module is now compatible with older versions of
containers. A small compatibility function needed to be created to
work around Set.alterF not existing in those versions.

The unsafePopChar function was removed and the functions that used it
were refactored. It really didn't need to exist when unsafeIndexSubtag
fills essentially the same role in those functions.

The Trie type was radically changed - it is now essentially isomorphic
to Maybe (NonEmptyTrie.Trie a) - and the NonEmptyTrie module was
added. This type structure has the benefit of enforcing the pruning of
empty Trie branches; the older type allowed them to accumulate without
careful maintenance. It also has the (extremely minor and fringe)
benefit of allowing a faster Foldable.null implementation.

A TrieSpec module was added, containing various property tests for the
tries and their functions.