Split up languagetags-gen, add tests for it, modify docs
The languagetags-gen package is now split up a little more sensibly,
instead of being one huge module. The manual parsing test that was in
the old Gen.hs was added as the lone test in a new test suite.

The README was edited slightly

Some comments were moved around in languagetags/test