Improve variant ordering in rendering, add tests, improve documentation
The ordering of variants in renderBCP47 is now slightly different (and
hopefully a little nicer), as the documentation of that function
describes.

The toSubtags function is fixed and now orders the variants as
renderBCP47 does. Fixing it happened to involve eliminating the
Internal.Registry module, moving its contents to Registry, and
slightly duplicating toSubtags in Registry.Redundant.

A couple of property tests were added related to renderBCP47 and
toSubtags (to test that they interact with each other and with parsing
and validation correctly). This uncovered a small rendering bug (not
adding the initial x- to a private use tag).

The README was updated with a small overview of the package.

Other bits of the documentation were expanded or otherwise improved.