Reorganize Canonicalization, fix some bugs, change Trie functions
Canonicalization now no longer differentiates between fixed and
unfixed warnings. This makes the implementation of many of the
functions much simpler.

As part of the Canonicalization refactoring, many of the types were
made simpler, and the entire Internal.BCP47.Canonicalization module
was removed, with some of its functions modified and moved to the
public module and the rest deleted. Some unit tests were added to
ensure that linting recognizes certain problems with particular tags.

During that refactoring, a couple of bugs were discovered and fixed,
and the property tests were modified to catch such bugs in
future. They were:

1. The subtag parsers would fail to parse their input as expected when
   that input had length exactly 9 and otherwise consisted of valid
   characters; a length check that should have been present in a
   branch of a case expression was not there. The property tests
   failed to catch this because, among other things, they didn't test
   that the parsers would fail properly on too-long input. This was
   only caught by chance during manual debugging related to (2).

2. The lookupLax function did not correctly update the cached that it
   brings with it as it goes deeper into the Trie; it would only
   update it when encountering a non-empty Node for the first time,
   and would otherwise not update. The lookupLax property tests were
   modified so that they now try to look up nodes that are on fairly
   well-populated branches (sometimes with non-empty ancestors or
   descendants, with the node sometimes present or not). It was lucky
   that this was caught at all, since the existing property test would
   only catch it if a random path (of length at least one) happened to
   be an extension of an already-existing branch (with at least two
   nodes) in a distinct random trie.

Various trie-related functions now match the ordering of inputs that
the containers package uses, so that, e.g., insert has the node to be
inserted as its first argument.