Relax bounds, fix small test bugs, add isSubtagChar
The bounds on base and template-haskell are relaxed to allow for
compilation with GHC 9.0.1.

Some of the case-insensitivity tests are now modified so that they
will not convert non-ASCII characters to ASCII when converting to
lower case. Their implementation before (using Text.toLower) was
incorrect, since the parsers should not actually be case insensitive
on all characters, but only case-insensitive on ASCII characters.

A new isSubtagChar function (added to help with fixing the
case-insensitivity tests) was also added.