Modify subtag errors, add new syntag parser
The PopSubtagError now no longer includes the remainder of the input.
This allows it to be included more easily in other errors - otherwise
the remainder-of-input will tend to be duplicated in broader error
types, or the constructors of PopSubtagError will need to be inlined.

The parsers popSubtagLenWith and popSubtagLen were written using the new
popSubtagWith to parse an initial subtag from a particular
stream (explicitly strict text for the latter). The SyntaxError' type
was updated so it could be used with these functions, and to account for
the changes to PopSubtagError.