Modify the subtag format, add more tests
The subtag format now no longer caches whether or not the subtag
contains letters or digits (or both).

A major new registry test has been added - it checks that for every
subtag or tag entry in the registry, we can parse, validate, then render
the entry and we will end up with exactly that entry at the end.