Clean some modules, add registry test
We need to ensure that there are no long primary language subtags in the
registry, otherwise some of the types as they are currently defined will
be able to represent invalid tags. This should suffice for now, but we
may need to reconsider the representation of some of the types in
question to rule this out permanently.