Update tests, canonicalization
The registry now violates the standard (section 3.1.4), which says that
if x has preferred value y and y is changed in the registry to have
preferred value z, then x must also be changed to have preferred value
z. This possibility was discussed in the past (according to the mailing
list), and it was decided then that section 2.2.2 (governing extlangs
and what their preferred-value can be) should take priority in this
scenario. Thus it is now possible to have such an x->y->z chain, but
only if x is an extlang (and so forcing y and z to be language subtags).
Because of this, the tests of registry well-formedness had to be
updated, and the canonicalization procedure had to be slightly tweaked
so that we check if an extlang's preferred-value language subtag itself
has a preferred value (and use that if it does have one).

We also now test against a few newer GHCs.