Update to 2023-08-02 registry
Also required some adjustment to the variant record renderer, and we
needed to disable a particular registry test as the property it tests no
longer holds. (Fortunately, the lack of that property doesn't seem to be
spec-breaking from a cursory look, and I don't think we rely on it at
all).